Added a placeholder to point the location of the barrier
The current code has the correct ordering of kernel calls that will
allow further optimization by overlaping the compute load on the device
and on the host with the data trasfer between them.

However, there are no performance benefits right now, since all
Read and Write calls are still in blocking mode.